chore(reorg): move long-text to new file
