+++ src/data-workspace/inputs/index.js
@@ -1,2 +1,3 @@
 export * from './file-inputs.js'
+export * from './long-text.js'
 export * from './inputs.js'
