+++ src/data-workspace/inputs/long-text.js
@@ -1,0 +1,32 @@
+import React from 'react'
+import { useField } from 'react-final-form'
+import styles from './inputs.module.css'
+import { InputPropTypes } from './utils.js'
+
+export const LongText = ({ name, syncData, onKeyDown, lastSyncedValue }) => {
+    const { input, meta } = useField(name, {
+        subscription: { value: true, dirty: true, valid: true },
+    })
+
+    const handleBlur = () => {
+        const { value } = input
+        const { dirty, valid } = meta
+        if (dirty && valid && value !== lastSyncedValue) {
+            syncData(value)
+        }
+    }
+
+    return (
+        <textarea
+            className={styles.longText}
+            rows="4"
+            {...input}
+            onBlur={(e) => {
+                handleBlur()
+                input.onBlur(e)
+            }}
+            onKeyDown={onKeyDown}
+        />
+    )
+}
+LongText.propTypes = InputPropTypes
